# docs/swagger.yaml
basePath: /
definitions:
...
  schemas.CardsByCollectionID:
    properties:
      answer:
        type: string
      id:
        type: integer
      question:
        type: string
    type: object
  schemas.CreateCardReq:
    properties:
      answer:
        type: string
      question:
        type: string
    required:
    - question
    type: object
  schemas.CreateCardResp:
    properties:
      answer:
        type: string
      collectionID:
        type: integer
      id:
        type: integer
      question:
        type: string
    type: object
...
  schemas.GetCardByCollectionIDResp:
    properties:
      cards:
        items:
          $ref: '#/definitions/schemas.CardsByCollectionID'
        type: array
      collectionID:
        type: integer
    type: object
...
  schemas.UpdateCardReq:
    properties:
      answer:
        type: string
      id:
        type: integer
      question:
        type: string
    required:
    - id
    type: object
  schemas.UpdateCardResp:
    properties:
      answer:
        type: string
      collection_id:
        type: integer
      id:
        type: integer
      question:
        type: string
    type: object
...
  util.ErrorResponse:
    properties:
      error:
        type: string
    type: object
info:
  contact: {}
  description: REST API for flashcard learning platform.
  title: Card Backend API
  version: "1.1"
paths:
...
  /cards/{cardID}/:
    delete:
      parameters:
      - description: card id
        in: path
        name: cardID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete card
      tags:
      - cards
    patch:
      consumes:
      - application/json
      parameters:
      - description: card id
        in: path
        name: cardID
        required: true
        type: integer
      - description: card payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.UpdateCardReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.UpdateCardResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update card
      tags:
      - cards
    put:
      consumes:
      - application/json
      parameters:
      - description: card id
        in: path
        name: cardID
        required: true
        type: integer
      - description: card payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.UpdateCardReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.UpdateCardResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update card
      tags:
      - cards
...
  /collections/{collectionID}/cards/:
    get:
      parameters:
      - description: collection id
        in: path
        name: collectionID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.GetCardByCollectionIDResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get cards by collection
      tags:
      - cards
    post:
      consumes:
      - application/json
      parameters:
      - description: collection id
        in: path
        name: collectionID
        required: true
        type: integer
      - description: card payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.CreateCardReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/schemas.CreateCardResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create card
      tags:
      - cards
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"